API: Remove NumPy Operators
In order to simplify the API, reduce maintainance, and reduce
confusion(?); there is now only one operator implementation that
uses CuPy. This decision was approved by @xiadong-yu @nikitinvv @slnsrydn
@tekinbicer

The consequence is that only Linux platforms are now supported (or
Windows platforms that have a new enough version of CuPy).
