--- src/tike/operators/numpy/__init__.py
@@ -1,24 +1,0 @@
-"""Module for operators utilizing the NumPy library.
-
-This is the reference operator library designed to be array agonstic. In
-theory, any library that implements the __array_function__ interface should be
-able to use this implementation as a base.
-"""
-
-from .convolution import *
-from .flow import *
-from .lamino import *
-from .operator import *
-from .propagation import *
-from .ptycho import *
-from .shift import *
-
-__all__ = (
-    'Convolution',
-    'Flow',
-    'Lamino',
-    'Operator',
-    'Propagation',
-    'Ptycho',
-    'Shift',
-)
